commit
To encourage good grades, Hermosa High School has decided to award each student a bookstore credit that is 10 times the student’s grade point average. In other words, a student with a 3.2 grade point average receives a $32.0 credit. Create a class called BookstoreCredit that prompts a student for a name and grade point average and then passes the values to a method called computeDiscount() that displays a descriptive message. The message uses the student’s name, echoes the grade point average, and computes and displays the credit.